requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.5
python-multipart==0.0.9
itsdangerous==2.2.0
starlette==0.39.2

playwright==1.55.0
sentry-sdk==2.39.0
python-dotenv==1.0.1

beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
